Solved player spawn and movement

+++ Assets/Scripts/LoadingScreen.cs
@@ -1,5 +1,4 @@
 using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
@@ -8,66 +7,66 @@ using System;
 
 public class LoadingScreen : MonoBehaviour
 {
-	// Get level loader progress bar
-	public Scrollbar progressbar_level;
-	public TextMeshProUGUI loading_status;
+    // Get level loader progress bar
+    public Scrollbar progressbar_level;
+    public TextMeshProUGUI loading_status;
 
-	// Minimum loading time variables
-	private float StartLoadingTime = 0f;
-	private bool Loading = true;
-	public float LoadingMinTime = 5; // most important, minimum time to load scene
-	AsyncOperation loadingOperation;
+    // Minimum loading time variables
+    private float StartLoadingTime = 0f;
+    private bool Loading = true;
+    public float LoadingMinTime = 5; // most important, minimum time to load scene
+    AsyncOperation loadingOperation;
 
-	// Start is called before the first frame update
-	void Start()
-	{
-		if (MainMenu.play_button_pressed == true)
-		{
-			StartCoroutine(LoadAsyncOperation());
-			MainMenu.play_button_pressed = false;
-		}
-	}
+    // Start is called before the first frame update
+    void Start()
+    {
+        if (MainMenu.play_button_pressed == true)
+        {
+            StartCoroutine(LoadAsyncOperation());
+            MainMenu.play_button_pressed = false;
+        }
+    }
 
-	 private IEnumerator LoadAsyncOperation()
-	 {
-		// For now only ArenaIndustrial
-		// In the future, the play button should select a random map index, and that index is public in MainMenu, being sent here to load
-		loadingOperation = SceneManager.LoadSceneAsync((int) MainMenu.arenaSelected);
+    private IEnumerator LoadAsyncOperation()
+    {
+        // For now only ArenaIndustrial
+        // In the future, the play button should select a random map index, and that index is public in MainMenu, being sent here to load
+        loadingOperation = SceneManager.LoadSceneAsync(MainMenu.arenaSelected);
 
-		// Stop it so we can put a minimum loading time
-		loadingOperation.allowSceneActivation = false;
-		Loading = true;
-		StartLoadingTime = Time.time;
+        // Stop it so we can put a minimum loading time
+        loadingOperation.allowSceneActivation = false;
+        Loading = true;
+        StartLoadingTime = Time.time;
 
-		// As long as the level didn't load, fill loading progress bar
-		while (loadingOperation.progress < 1)
-		{
-			// this is the progress bar of level load
-			// will be, for this game, always up to 90%, because it just loads so fast due to the small game
-			// but we will limit by time
-			progressbar_level.value = (float) loadingOperation.progress;
-			yield return new WaitForEndOfFrame();
-		}
+        // As long as the level didn't load, fill loading progress bar
+        while (loadingOperation.progress < 1)
+        {
+            // this is the progress bar of level load
+            // will be, for this game, always up to 90%, because it just loads so fast due to the small game
+            // but we will limit by time
+            progressbar_level.value = (float)loadingOperation.progress;
+            yield return new WaitForEndOfFrame();
+        }
 
-		yield return new WaitForEndOfFrame();
-	 }
+        yield return new WaitForEndOfFrame();
+    }
 
-	private void Update()
-	{
-		if (Loading)
-		{
-			float loadingTime = Time.time - StartLoadingTime;
-			if((Math.Floor((LoadingMinTime - loadingTime) * 100) / 100) > 0)
-				loading_status.text = "Game loaded! Prepare to fight in " + (Math.Floor((LoadingMinTime - loadingTime)  * 100) / 100).ToString();
+    private void Update()
+    {
+        if (Loading)
+        {
+            float loadingTime = Time.time - StartLoadingTime;
+            if ((Math.Floor((LoadingMinTime - loadingTime) * 100) / 100) > 0)
+                loading_status.text = "Game loaded! Prepare to fight in " + (Math.Floor((LoadingMinTime - loadingTime) * 100) / 100).ToString();
 
-			// Checks if the min loading time has passed
-			if (loadingTime >= LoadingMinTime)
-			{
-				// At this point if the loading is really over, the scene will load
-				// If the load it's not over, it will continue to load and switch to the scene at the end
-				loadingOperation.allowSceneActivation = true;
-				Loading = false;
-			}
-		}
-	}
+            // Checks if the min loading time has passed
+            if (loadingTime >= LoadingMinTime)
+            {
+                // At this point if the loading is really over, the scene will load
+                // If the load it's not over, it will continue to load and switch to the scene at the end
+                loadingOperation.allowSceneActivation = true;
+                Loading = false;
+            }
+        }
+    }
 }
